feat: reorganizar importaciones y crear nuevas interfaces para la gestión de campañas y enciclopedia cumpliendo con arquitectura FSD
