+++ src/views/campaigns/index.ts
@@ -1,0 +1,6 @@
+// Public API — campaigns slice
+export { getUserCampaigns, getCampaignById } from './api/campaign-queries'
+export { createCampaign, updateCampaign } from './api/campaign-actions'
+export type { ActionResponse, CreateCampaignDTO, UpdateCampaignDTO } from './api/campaign-actions'
+export { default as CampaignCreationProvider } from './ui/creation/CampaignCreationProvider'
+export type { CampaignFormValues } from './model/campaign'
