+++ src/views/campaigns/model/campaign.ts
