+++ src/views/encyclopedia/model/encyclopedia.ts
