+++ src/views/encyclopedia/ui/DetailView.tsx
@@ -4,7 +4,7 @@ import React, { useState } from 'react'
 import { m, AnimatePresence } from 'framer-motion'
 import { Info, Shield, Heart, Activity } from 'lucide-react'
 import { type EncyclopediaSection } from '../model/encyclopediaStore'
-import { EncyclopediaItem } from '../model/types'
+import { EncyclopediaItem } from '../model/encyclopedia'
 import Image from 'next/image'
 import { cn } from '@/shared/utils/styles'
 import { getEntityImage } from '@/shared/lib/storage'
