+++ src/views/encyclopedia/ui/EncyclopediaStoreInitializer.tsx
@@ -2,7 +2,7 @@
 
 import { useRef } from 'react'
 import { useEncyclopediaStore, type EncyclopediaSection } from '../model/encyclopediaStore'
-import { EncyclopediaItem } from '../model/types'
+import { EncyclopediaItem } from '../model/encyclopedia'
 
 interface Props {
   data: Record<EncyclopediaSection, EncyclopediaItem[]>
@@ -10,7 +10,7 @@ interface Props {
 
 export function EncyclopediaStoreInitializer({ data }: Props) {
   const initialized = useRef(false)
-  
+
   if (!initialized.current) {
     useEncyclopediaStore.getState().setItems(data)
     initialized.current = true
