+++ src/views/portal/index.ts
@@ -1,0 +1,4 @@
+// Public API — portal slice
+export { PortalCarousel } from './ui/portal-carousel'
+export { PortalCard } from './ui/portal-card'
+export { Portal } from './ui/Portal'
